REFACTOR model.js into two files
